turn of actors after specific time

+++ app/src/main/java/com/bach/gardenstate/features/actors/model/PoolPumpData.kt
@@ -7,7 +7,7 @@ data class PoolPumpData(
     var state: String,
     val power: Int,
     val voltage: Int,
-    val current: Int,
+    val current: Float,
     val energy: Float,
     val linkquality: Int,
     val last_seen: String
